+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.8.5
+
+## Bug fixes
+
+- Fixed an issue where `pulsar_producers` would stop and not restart if it encountered
+  problems when trying to reach `pulsar_client`.
+
 # 0.8.3
 
 ## Bug fixes